+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 Changelog
 =========
 
+## Unreleased
+### Added
+- `flipbox\element\lists\db\SourceUserElementQuery` to build a full query for users
+
+### Changed
+- Fields can provide their own query class
+
 ## 1.0.0 - 2018-04-30
 - Initial release